+++ Xext/xselinux/meson.build
@@ -1,0 +1,11 @@
+srcs_xselinux = [
+    'xselinux_ext.c',
+    'xselinux_hooks.c',
+    'xselinux_label.c',
+]
+
+libxserver_xselinux = static_library('xserver_xselinux',
+    srcs_xselinux,
+    include_directories: inc,
+    dependencies: common_dep,
+)
